+++ app/ui/src/pages/System/index.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import { useTranslation } from 'react-i18next';
 import { SystemMonitor } from './SystemMonitor';
+import { ConfigAuditCard } from './ConfigAuditCard';
 import { ProcessorLogs } from './ProcessorLogs';
 import { StorageOverview } from './StorageOverview';
 import { Box, Divider } from '@mui/material';
@@ -14,6 +15,12 @@ export const System: React.FC = () => {
       <Box>
         <Box>
           <SystemMonitor />
+        </Box>
+
+        <Divider sx={{ my: 3 }} />
+
+        <Box>
+          <ConfigAuditCard />
         </Box>
 
         <Divider sx={{ my: 3 }} />